fix: add back ExpectedAnswerCard.tsx
