+++ src/requirements/Form/components/types.ts
@@ -1,0 +1,7 @@
+export type ExpectedFieldDataProps = {
+  value?: string
+  minAmount?: number
+  maxAmount?: number
+  acceptedAnswers?: string[]
+  rejectedAnswers?: string[]
+}
